+++ src/SendGrid.php
@@ -1,7 +1,9 @@
 <?php
+
 namespace Sichikawa\LaravelSendgridDriver;
 
 use Illuminate\Mail\Mailable;
+use Sichikawa\LaravelSendgridDriver\Transport\SendgridTransport;
 use Swift_Message;
 
 trait SendGrid
@@ -13,11 +15,19 @@ trait SendGrid
      */
     public function sendgrid($params)
     {
-        if ($this instanceof Mailable && env('MAIL_DRIVER') == "sendgrid") {
+        if ($this instanceof Mailable && $this->mailDriver() == "sendgrid") {
             $this->withSwiftMessage(function (Swift_Message $message) use ($params) {
-                $message->embed(\Swift_Image::newInstance($params, 'sendgrid/x-smtpapi'));
+                $message->embed(\Swift_Image::newInstance($params, SendgridTransport::SMTP_API_NAME));
             });
         }
         return $this;
+    }
+
+    /**
+     * @return string
+     */
+    private function mailDriver()
+    {
+        return function_exists('config') ? config('mail.driver') : env('MAIL_DRIVER');
     }
 }
