+++ src/SendgridTransportServiceProvider.php
@@ -6,7 +6,6 @@ use Illuminate\Mail\TransportManager;
 use Illuminate\Support\Arr;
 use Illuminate\Support\ServiceProvider;
 use Sichikawa\LaravelSendgridDriver\Transport\SendgridTransport;
-use Sichikawa\LaravelSendgridDriver\Transport\SendgridV3Transport;
 
 class SendgridTransportServiceProvider extends ServiceProvider
 {
@@ -27,10 +26,6 @@ class SendgridTransportServiceProvider extends ServiceProvider
         $manager->extend('sendgrid', function() {
             $config = $this->app['config']->get('services.sendgrid', array());
             $client = new HttpClient(Arr::get($config, 'guzzle', []));
-
-            if (Arr::get($config, 'version') === 'v3') {
-                return new SendgridV3Transport($client, $config['api_key']);
-            }
 
             return new SendgridTransport($client, $config['api_key']);
         });
