+++ tests/SendgridTransportServiceProviderTest.php
@@ -3,20 +3,15 @@
 use Illuminate\Mail\TransportManager;
 use Sichikawa\LaravelSendgridDriver\SendgridTransportServiceProvider;
 use Sichikawa\LaravelSendgridDriver\Transport\SendgridTransport;
-use Sichikawa\LaravelSendgridDriver\Transport\SendgridV3Transport;
 
 class SendgridTransportServiceProviderTest extends TestCase
 {
-    public function registerServiceProvider($version = null)
+    public function registerServiceProvider()
     {
         $app = $this->getMockForAbstractClass(\Illuminate\Container\Container::class);
 
         $app['config'] = new \MockConfig();
         $app['config']->set('mail.driver', 'sendgrid');
-
-        if ($version) {
-            $app['config']->set('services.sendgrid.version', $version);
-        }
 
         $serviceProvider = new SendgridTransportServiceProvider($app);
         $transportManager = new TransportManager($app);
@@ -27,17 +22,10 @@ class SendgridTransportServiceProviderTest extends TestCase
 
     public function testCreateSendgridDriver()
     {
-        $transportManager = $this->registerServiceProvider('v2');
+        $transportManager = $this->registerServiceProvider();
+
         $sendgridDriver = $transportManager->driver();
+
         $this->assertInstanceOf(SendgridTransport::class, $sendgridDriver);
-    }
-
-    public function testCreateSendgridV3Driver()
-    {
-        $transportManager = $this->registerServiceProvider('v3');
-
-        $sendgridDriver = $transportManager->driver();
-
-        $this->assertInstanceOf(SendgridV3Transport::class, $sendgridDriver);
     }
 }
