+++ tests/Transport/SendgridTransportTest.php
@@ -1,8 +1,8 @@
 <?php
 
-use Sichikawa\LaravelSendgridDriver\Transport\SendgridTransport;
 use GuzzleHttp\Client as HttpClient;
 use Illuminate\Mail\Message;
+use Sichikawa\LaravelSendgridDriver\Transport\SendgridTransport;
 
 class SendgridTransportTest extends TestCase
 {
@@ -25,138 +25,242 @@ class SendgridTransportTest extends TestCase
         $message->from('from@sink.sendgrid.net', 'test_from')
             ->to('to@sink.sendgrid.net', 'test_to');
         $res = $this->transport->send($message->getSwiftMessage());
-        $this->assertEquals(200, $res->getStatusCode());
+        $this->assertEquals(1, $res);
+    }
+
+    public function testXMessageID()
+    {
+        $client = $this->getMockBuilder(\GuzzleHttp\Client::class)
+            ->setMethods(['post'])
+            ->getMock();
+        $transport = new SendgridTransport($client, $this->api_key);
+
+        $messageId = str_random(32);
+        $client->expects($this->once())
+            ->method('post')
+            ->willReturn(new \GuzzleHttp\Psr7\Response(200, [
+                'X-Message-ID' => $messageId
+            ]));
+        $message = new Message($this->getMessage());
+        $message->from('from@sink.sendgrid.net', 'test_from')
+            ->to('to@sink.sendgrid.net', 'test_to');
+        $transport->send($message->getSwiftMessage());
+        $this->assertEquals($messageId, $message->getSwiftMessage()->getHeaders()->get('X-Message-ID')->getFieldBody());
     }
 
     public function testMultipleSend()
     {
         $message = new Message($this->getMessage());
         $message->from('from@sink.sendgrid.net', 'test_from')
-            ->to('foo@sink.sendgrid.net', 'foo');
+            ->to('foo@sink.sendgrid.net', 'foo')
+            ->cc('foo2@sink.sendgrid.net', 'foo2');
         $res = $this->transport->send($message->getSwiftMessage());
 
-        $this->assertEquals(200, $res->getStatusCode());
+        $this->assertEquals(2, $res);
 
         $message = new Message($this->getMessage());
         $message->from('from@sink.sendgrid.net', 'test_from')
-            ->to('bar@sink.sendgrid.net', 'bar');
+            ->to('bar@sink.sendgrid.net', 'bar')
+            ->bcc('bar2@sink.sendgrid.net', 'bar2');
         $res = $this->transport->send($message->getSwiftMessage());
 
-        $this->assertEquals(200, $res->getStatusCode());
+        $this->assertEquals(2, $res);
     }
 
-    public function testSetTo()
+    public function testSendByPersonalization()
     {
-        $to = 'test@exsample.com';
-        $to_name = 'test_user';
-        $setTo = \Closure::bind(function (&$data, $message) {
-            $this->setTo($data, $message);
-        }, $this->transport, 'Sichikawa\LaravelSendgridDriver\Transport\SendGridTransport');
-        $data = [];
+        $message = new Message($this->getMessage());
+        $message->from('from@sink.sendgrid.net', 'test_from')
+            ->to('dummy@sink.sendgrid.net')
+            ->embedData([
+                'personalizations' => [
+                    [
+                        'to' => [
+                            'email' => 'foo@sink.sendgrid.net',
+                            'name'  => 'foo',
+                        ],
+                    ],
+                ],
+            ], 'sendgrid/x-smtpapi');
+        $res = $this->transport->send($message->getSwiftMessage());
+
+        $this->assertEquals(1, $res);
+
+        $message = new Message($this->getMessage());
+        $message->from('from@sink.sendgrid.net', 'test_from')
+            ->to('dummy@sink.sendgrid.net')
+            ->embedData([
+                'personalizations' => [
+                    [
+                        'to' => [
+                            'email' => 'to1@sink.sendgrid.net',
+                            'name'  => 'to1',
+                        ],
+                    ],
+                    [
+                        'to' => [
+                            'email' => 'to2@sink.sendgrid.net',
+                            'name'  => 'to2',
+                        ],
+                        'cc' => [
+                            'email' => 'cc1@sink.sendgrid.net',
+                            'name'  => 'cc1',
+                        ]
+                    ],
+                    [
+                        'to' => [
+                            'email' => 'to3@sink.sendgrid.net',
+                            'name'  => 'to3',
+                        ],
+                        'bcc' => [
+                            'email' => 'bcc1@sink.sendgrid.net',
+                            'name'  => 'bcc1',
+                        ]
+                    ],
+                ],
+            ], 'sendgrid/x-smtpapi');
+        $res = $this->transport->send($message->getSwiftMessage());
+
+        $this->assertEquals(5, $res);
+    }
+
+    public function testGetPersonalizations()
+    {
+
+
+        $getPersonalizations = \Closure::bind(function ($message) {
+            return $this->getPersonalizations($message);
+        }, $this->transport, SendgridTransport::class);
+
         $message = $this->getMessage();
+
+        $to = 'to_user@exsample.com';
+        $to_name = 'to_user';
         $message->setTo($to, $to_name);
-        $setTo($data, $message);
-        $this->assertEquals($to, array_pop($data['to']));
-        $this->assertEquals($to_name, array_pop($data['toname']));
-    }
 
-    public function testSetCc()
-    {
-        $cc = 'test@exsample.com';
-        $cc_name = 'test_user';
-        $setCc = \Closure::bind(function (&$data, $message) {
-            $this->setCc($data, $message);
-        }, $this->transport, 'Sichikawa\LaravelSendgridDriver\Transport\SendGridTransport');
-        $data = [];
-        $message = $this->getMessage();
+        $cc = 'cc_user@exsample.com';
+        $cc_name = 'cc_user';
         $message->setCc($cc, $cc_name);
-        $setCc($data, $message);
-        $this->assertEquals($cc, array_pop($data['cc']));
-        $this->assertEquals($cc_name, array_pop($data['ccname']));
-    }
 
-    public function testSetBcc()
-    {
-        $bcc = 'test@exsample.com';
-        $bcc_name = 'test_user';
-        $setBcc = \Closure::bind(function (&$data, $message) {
-            $this->setBcc($data, $message);
-        }, $this->transport, 'Sichikawa\LaravelSendgridDriver\Transport\SendGridTransport');
-        $data = [];
-        $message = $this->getMessage();
+        $bcc = 'bcc_user@exsample.com';
+        $bcc_name = 'bcc_user';
         $message->setBcc($bcc, $bcc_name);
-        $setBcc($data, $message);
-        $this->assertEquals($bcc, array_pop($data['bcc']));
-        $this->assertEquals($bcc_name, array_pop($data['bccname']));
+
+        $res = $getPersonalizations($message);
+
+        $this->assertEquals($to, array_get($res, '0.to.0.email'));
+        $this->assertEquals($to_name, array_get($res, '0.to.0.name'));
+        $this->assertEquals($cc, array_get($res, '0.cc.0.email'));
+        $this->assertEquals($cc_name, array_get($res, '0.cc.0.name'));
+        $this->assertEquals($bcc, array_get($res, '0.bcc.0.email'));
+        $this->assertEquals($bcc_name, array_get($res, '0.bcc.0.name'));
     }
 
-    public function testSetText()
+    public function testGetContents()
     {
-        $setText = \Closure::bind(function (&$data, $message) {
-            return $this->setText($data, $message);
-        }, $this->transport, 'Sichikawa\LaravelSendgridDriver\Transport\SendGridTransport');
+        $getContents = \Closure::bind(function ($message) {
+            return $this->getContents($message);
+        }, $this->transport, SendgridTransport::class);
 
-        $data = [];
         $message = new Message($this->getMessage());
         $message->getSwiftMessage()->setChildren([Swift_MimePart::newInstance(
             'This is a test.'
         )]);
-        $setText($data, $message->getSwiftMessage());
 
-        $this->assertEquals('This is a test.', $data['text']);
+        $res = $getContents($message->getSwiftMessage());
+        $this->assertEquals('text/plain', array_get($res, '0.type'));
+        $this->assertEquals('This is a test.', array_get($res, '0.value'));
+        $this->assertEquals('text/html', array_get($res, '1.type'));
+        $this->assertEquals('Test body.', array_get($res, '1.value'));
     }
 
-    public function testSetAttachment()
+    public function testGetAttachments()
     {
-        $setAttachment = \Closure::bind(function (&$data, $message) {
-            return $this->setAttachment($data, $message);
-        }, $this->transport, 'Sichikawa\LaravelSendgridDriver\Transport\SendGridTransport');
-        $data = [];
+        $getAttachment = \Closure::bind(function ($message) {
+            return $this->getAttachments($message);
+        }, $this->transport, SendgridTransport::class);
+
         $message = new Message($this->getMessage());
-        $message->attach(__DIR__ . '/test.png');
-        $setAttachment($data, $message->getSwiftMessage());
-        $this->assertEquals('stream', get_resource_type($data['files[test.png]']));
+
+        $file = __DIR__ . '/test.png';
+        $message->attach($file);
+
+        $res = $getAttachment($message->getSwiftMessage());
+        $this->assertEquals(base64_encode(file_get_contents($file)), array_get($res, '0.content'));
+        $this->assertEquals('test.png', array_get($res, '0.filename'));
     }
 
-    public function testSetSmtpApi()
+    public function testSetParameters()
     {
-        $setSmtpApi = \Closure::bind(function (&$data, $message) {
-            return $this->setSmtpApi($data, $message);
-        }, $this->transport, 'Sichikawa\LaravelSendgridDriver\Transport\SendGridTransport');
+        $setParameters = \Closure::bind(function ($message, $data) {
+            return $this->setParameters($message, $data);
+        }, $this->transport, SendgridTransport::class);
+
+        $parameters = [
+            'categories' => 'category1',
+        ];
+        $message = new Message($this->getMessage());
+        $message->embedData($parameters, 'sendgrid/x-smtpapi');
         $data = [];
+        $data = $setParameters($message->getSwiftMessage(), $data);
+        $this->assertEquals($parameters, $data);
+    }
+
+    public function testSetPersonalizations()
+    {
+        $setParameters = \Closure::bind(function ($message, $data) {
+            return $this->setParameters($message, $data);
+        }, $this->transport, SendgridTransport::class);
+
+        $personalizations = [
+            [
+                'substitutions' => [
+                    'substitutions_key' => 'substitutions_value',
+                ],
+                'custom_args'   => [
+                    'custom_args_key' => 'custom_args_value',
+                ],
+                'send_at'       => time(),
+            ],
+        ];
+
         $message = new Message($this->getMessage());
         $message->embedData([
-            'category' => 'foo'
+            'personalizations' => $personalizations,
         ], 'sendgrid/x-smtpapi');
-        $setSmtpApi($data, $message->getSwiftMessage());
-        $this->assertEquals(json_encode(['category' => 'foo']), $data['x-smtpapi']);
+        $data = [];
+        $data = $setParameters($message->getSwiftMessage(), $data);
+        $this->assertEquals(['personalizations' => $personalizations], $data);
     }
 
-    public function testGetFromAddresses()
+    public function testGetFrom()
     {
+        $getFrom = \Closure::bind(function ($message) {
+            return $this->getFrom($message);
+        }, $this->transport, SendgridTransport::class);
+
         $from = 'test@exsample.com';
         $from_name = 'test_user';
-        $getFromAddresses = \Closure::bind(function ($message) {
-            return $this->getFromAddresses($message);
-        }, $this->transport, 'Sichikawa\LaravelSendgridDriver\Transport\SendGridTransport');
+
         $message = $this->getMessage();
         $message->setFrom($from, $from_name);
-        $this->assertEquals([$from, $from_name], $getFromAddresses($message));
+
+        $this->assertEquals(['email' => $from, 'name' => $from_name], $getFrom($message));
     }
 
-
-    public function testSetReplyTo()
+    public function testReplyTo()
     {
+        $getReplyTo = \Closure::bind(function ($message) {
+            return $this->getReplyTo($message);
+        }, $this->transport, SendgridTransport::class);
+
         $reply_to = 'test@exsample.com';
         $reply_to_name = 'test_user';
-        $setReplyTo = \Closure::bind(function (&$data, $message) {
-            $this->setReplyTo($data, $message);
-        }, $this->transport, 'Sichikawa\LaravelSendgridDriver\Transport\SendGridTransport');
-        $data = [];
+
         $message = $this->getMessage();
         $message->setReplyTo($reply_to, $reply_to_name);
-        $setReplyTo($data, $message);
-        $this->assertEquals($reply_to, $data['replyto']);
+
+        $this->assertEquals(['email' => $reply_to, 'name' => $reply_to_name], $getReplyTo($message));
     }
 
     /**
