+++ tests/config/services.php
@@ -1,6 +1,6 @@
 <?php
 return [
     'sendgrid' => [
-        'api_key' => ''
+        'api_key' => env('SENDGRID_API_KEY')
     ]
 ];
